requirements.txt
# Lead Generation Project Dependencies

# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Web dashboard
streamlit>=1.30.0
pandas>=2.0.0
